+++ .github/workflows/ci-workflow-server.yaml
@@ -1,33 +1,30 @@
 name: ci-workflow-server
 
-on:
-  push:
-    branches:
-      - main
+on: [push]
+#on:
+#  push:
+#    branches:
+#      - main
 jobs:
   publish:
     permissions:
       contents: "read"
       id-token: "write"
 
-    if: github.event_name == 'push' && contains(github.ref, 'refs/tags/')
+#    if: github.event_name == 'push' && contains(github.ref, 'refs/tags/')
     runs-on: ubuntu-20.04
     steps:
       - name: Checkout repo
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
 
       - name: Set up Node.js
         uses: actions/setup-node@v3
         with:
-          node-version: '18.17.0'
+          node-version: "18.18.0"
 
       - name: Install dependencies
         run: npm install
-        working-directory: ee/codegen
-
-      - name: Build
-        run: npm run build
-        working-directory: ee/codegen
+        working-directory: ee/automation
 
       - id: auth
         uses: google-github-actions/auth@v2
@@ -44,4 +41,4 @@ jobs:
           VELLUM_AUTOMATION_PRIVATE_KEY: ${{ secrets.VELLUM_AUTOMATION_PRIVATE_KEY }}
           VELLUM_AUTOMATION_INSTALLATION_ID: ${{ secrets.VELLUM_AUTOMATION_INSTALLATION_ID }}
         run: npm run upgrade-workflow-server
-        working-directory: ee/codegen
+        working-directory: ee/automation
